Package Handling.
We're reworking the way that information is shared across the system.
Now a dictionary is being used as a packet through the use of base64.
Using the JSON module we can convert a properly formatted string into
a dictionary-like object that can be used to silo access to specific
pieces of information included with the message.