New Imports.
Some things are being introduced; such as base64, and logging.  Base64
allows for encoding objects as bytes, while logging is for logging.

The specific members of logging being used are exception, info, and
debug. All called individually and renamed in the long form except for
exception, which speaks for itself.